+++ app/routes.py
@@ -30,8 +30,15 @@ from datetime import datetime
 import re
 
 from app.cache_utils import get_cache, set_cache
-from app.models import Player, Statistics, LeagueDashPlayerStats, Team,PlayerGameLog, get_player_data
-from app.utils import get_todays_games_and_standings, get_enhanced_teams_data, get_team_lineup_stats
+
+from app.models.player import Player
+from app.models.statistics import Statistics
+from app.models.team import Team
+from app.models.leaguedashplayerstats import LeagueDashPlayerStats
+from app.models.playergamelog import PlayerGameLog
+from app.models.gameschedule import GameSchedule
+
+from app.utils import get_todays_games_and_standings, get_enhanced_teams_data, get_team_lineup_stats, get_player_data
 
 main = Blueprint("main", __name__)
